Complete the move of null-loader to core
