+++ modules/core/src/index.ts
@@ -29,7 +29,7 @@ export {RequestScheduler} from '@loaders.gl/loader-utils';
 export {makeIterator} from './iterator-utils/make-iterator/make-iterator';
 
 // CORE LOADERS
-export {NullWorkerLoader, NullLoader} from '@loaders.gl/loader-utils';
+export {NullWorkerLoader, NullLoader} from './null-loader';
 export {JSONLoader} from '@loaders.gl/loader-utils';
 
 // EXPERIMENTAL
