+++ modules/core/src/null-loader.ts
@@ -2,7 +2,24 @@
 // @ts-ignore TS2304: Cannot find name '__VERSION__'.
 const VERSION = typeof __VERSION__ !== 'undefined' ? __VERSION__ : 'latest';
 
-import {LoaderWithParser} from '@loaders.gl/loader-utils';
+import {Loader, LoaderWithParser} from '@loaders.gl/loader-utils';
+
+/**
+ * Loads any data and returns null (or optionally passes through data unparsed)
+ */
+export const NullWorkerLoader: Loader = {
+  name: 'Null loader',
+  id: 'null',
+  module: 'core',
+  version: VERSION,
+  worker: true,
+  mimeTypes: ['application/x.empty'],
+  extensions: ['null'],
+  tests: [() => false],
+  options: {
+    null: {}
+  }
+};
 
 /**
  * Loads any data and returns null (or optionally passes through data unparsed)
