+++ modules/loader-utils/package.json
@@ -42,6 +42,7 @@
   "dependencies": {
     "@babel/runtime": "^7.3.1",
     "@loaders.gl/worker-utils": "3.0.0-beta.1",
+    "@loaders.gl/core": "3.0.0-beta.1",
     "@probe.gl/stats": "^3.3.0"
   }
 }
