+++ modules/loader-utils/src/workers/null-worker.ts
@@ -1,4 +1,4 @@
 import {createLoaderWorker} from '../lib/worker-loader-utils/create-loader-worker';
-import {NullLoader} from '../null-loader';
+import {NullLoader} from '@loaders.gl/core';
 
 createLoaderWorker(NullLoader);
